Template makeFile generated upon use of Submit button.
